// src/app/dashboard/model/admin.ts
export class Admin {
    subOperatorId: number;
    operatorId: number;
    subOperatorName: string;
    email: string;
    contactPerson: string;
    contactNumber: number;
    phone: number;
    location: string;
    address: string;
    operatorType: string;
    mainSubOperatorId: number;
    userID: number;
    blocked: boolean;
    enabled: boolean;
    freeTrail: boolean;
    billDate: number;
    paymentDate: number;
    freeTrailEndDate: number;
    createdAt: number;
    consentUpdateAt: number;
    changedAt: number;
    userName: string;
}